+++ sdk/python/foundation-models/system/content-safety/safety-llama-test/sample-request.json
@@ -1,0 +1,1 @@
+{"input_data": {"data": ["Hello", "My name is John and I have a dog.", "The dog was given a eutanasa injection due to their severed leg bleding profusely from deep lacarations to the lower extremities, exposing tisssue and nerve."]}}
