+++ sdk/python/foundation-models/system/inference/text-generation/llama-files/docker_env/Dockerfile
